requirements.txt
Flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-login>=0.6.3
flask-socketio>=5.3.6
python-socketio>=5.10.0
APScheduler>=3.10.4
groq>=0.4.2
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
sqlalchemy>=2.0.23
werkzeug>=3.0.1
numpy>=1.26.0
gunicorn>=21.2.0
eventlet>=0.33.3